requirements.txt
fastapi==0.61.1
uvicorn==0.11.8
google-cloud-speech==2.11.1